fix calcDeltasToCLoseSetRatio function

+++ test/foundry/CalcDeltasForMaxLP.t.sol
@@ -21,7 +21,7 @@ contract CalcDeltasForMaxLP is CalcDeltasBase {
         }
 
         (tokensHeld0, tokensHeld1, reserve0, reserve1) = updateTokenQtys(tokensHeld0, tokensHeld1, reserve0, reserve1,
-            uint256(deltas[0]), uint256(deltas[1]));
+            deltas[0], deltas[1]);
 
         if(deltas[0] > 0 || deltas[1] > 0) {
             assertGe(GSMath.sqrt(uint256(tokensHeld0) * tokensHeld1), collateral0); // when rebalancing is small it might not pass
