+++ test/foundry/CalcDeltasForWithdrawal.t.sol
@@ -20,7 +20,7 @@ contract CalcDeltasForWithdrawal is CalcDeltasBase {
         } // otherwise no trade
 
         (tokensHeld0, tokensHeld1, reserve0, reserve1) = updateTokenQtys(tokensHeld0, tokensHeld1, reserve0, reserve1,
-            uint256(deltas[0]), uint256(deltas[1]));
+            deltas[0], deltas[1]);
 
         if(deltas[0] > 0) {
             tokensHeld0 -= uint128(amounts[0]);
